Redux

+++ src/components/MyCard/MyCard.tsx
@@ -8,15 +8,17 @@ import {
 } from "@material-ui/core";
 import {makeStyles} from "@material-ui/core/styles";
 import ICardsDataDTO from "../../models/ICardsDataDTO";
-import {myCardAction} from "../../models/myCardAction";
+import {myCardAction} from "../../models/enumsActions/myCardAction";
+import cardReducer from "../../redux/reducers/cardReducer";
 
 interface ICardProps {
     headerText: string;
     bodyText: string;
     id: string;
-    setCards?: React.Dispatch<React.SetStateAction<ICardsDataDTO[]>>;
-    globalIsEdit: boolean;
+    setCards?: (cards: ICardsDataDTO[]) => void
+    isEdit: boolean;
     isSave: boolean;
+    handleDeleteBtnClick?: (id: string) => void
 }
 
 const useStyles = makeStyles({
@@ -51,55 +53,17 @@ const useStyles = makeStyles({
     },
 });
 
-interface ICardState {
-    title?: string;
-    body?: string;
-    prevTitle?: string;
-    prevBody?: string;
-    isErrorTitle?: boolean;
-    isErrorBody?: boolean;
-}
-
-const cardReducer = (
-    state: ICardState,
-    action: { type: string; payload: ICardState }
-): ICardState => {
-    switch (action.type) {
-        case myCardAction.TITLE_CHANGE: {
-            const newState = {...state};
-            newState.title = action.payload.title;
-            newState.isErrorTitle = action.payload.isErrorTitle;
-            return newState;
-        }
-        case myCardAction.BODY_CHANGE: {
-            const newState = {...state};
-            newState.body = action.payload.body;
-            newState.isErrorBody = action.payload.isErrorBody;
-            return newState;
-        }
-        case myCardAction.PREV_TITLE_CHANGE: {
-            const newState = {...state};
-            newState.prevTitle = action.payload.prevTitle;
-            return newState;
-        }
-        case myCardAction.PREV_BODY_CHANGE: {
-            const newState = {...state};
-            newState.prevBody = action.payload.prevBody;
-            return newState;
-        }
-        default:
-            return state;
-    }
-};
 
 const MyCard = ({
                     headerText = "title",
                     bodyText = "body",
                     id = "0",
                     setCards,
-                    globalIsEdit,
+                    isEdit,
                     isSave,
+                    handleDeleteBtnClick
                 }: ICardProps) => {
+
     const defaultValue = {
         title: headerText,
         body: bodyText,
@@ -135,15 +99,7 @@ const MyCard = ({
             }
         };
         check();
-    }, [isSave]);
-
-    const handleDeleteBtnClick = (id: string) => {
-        if (setCards) {
-            setCards((prevState) => {
-                return prevState.filter((item) => item.id !== id);
-            });
-        }
-    };
+    }, [isEdit]);
 
     const handleTitle = (e: ChangeEvent<HTMLInputElement>) => {
         cardDispatch({
@@ -166,11 +122,11 @@ const MyCard = ({
 
     return (
         <Card className={classes.card}>
-            {globalIsEdit && setCards && (
+            {isEdit && setCards && (
                 <button
                     className={classes.cardBtn}
                     onClick={() => {
-                        handleDeleteBtnClick(id);
+                        handleDeleteBtnClick && handleDeleteBtnClick(id);
                     }}
                 >
                     <span className="material-icons">clear</span>
@@ -178,7 +134,7 @@ const MyCard = ({
             )}
             <CardActionArea>
                 <CardContent>
-                    {globalIsEdit ? (
+                    {isEdit ? (
                         <div>
                             <TextField
                                 required
@@ -197,7 +153,7 @@ const MyCard = ({
                         </Typography>
                     )}
 
-                    {globalIsEdit ? (
+                    {isEdit ? (
                         <div>
                             <TextField
                                 required
